pyproject.toml
[project]
name = "quintessencia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "neonize>=0.3.16.post0",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
]
